requirements.txt
Django>=3.0.6,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.5,<2.9.0 #django communicating with postgres
Pillow>=5.3.0,<5.4.0